Revert "new nav bar attempt"
