+++ src/Front-End/Pages/CoursePage/CoursePage.css
@@ -1,6 +1,6 @@
 .SearchTest {
     width: 100vw;
-    min-height: 200vh;
+    min-height: 100vh;
     background-color: rgb(230, 213, 251);
     background-size: cover;
 }
@@ -48,8 +48,9 @@
     flex-direction: column;
     justify-content: left;
     top: 100%;
-    left: 10%;
-    width: 80%;
+    left: 0%;
+    width: 45%;
+    height: 35%
 }
 
 #143grade {
@@ -77,8 +78,8 @@
     align-items: left;
     flex-direction: column;
     justify-content: left;
-    top: 350%;
-    left: 20%;
+    top: 100%;
+    left: 50%;
 }
 
 .Reviews > div {
